1)Decomposed health_declarations
2)Merged contacts back with Employees. Doesn’t break any normalization rules and reduces the number of tables

3)Added how to handle deletes and updates

For anything that involves eid, use ON UPDATE CASCADE. Default ON DELETE NO ACTION is used for anything that references eid since we keep records of employees even after they resign. (ie, employees are only soft deleted by adding a resignation date on them)

If the room changes to a different place (room/floor), all updates for that room should be deleted first since by definition, the room is different. Hence, ON DELETE NO ACTION

If a meeting room is deleted, all updates and sessions should be deleted as well.

If a session is deleted, Approves tuple and Joins tuples should be deleted as well